+++ Universe/Assets/Animations/Monsters/Boss/Atattack01.controller.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: a6e89906177905c4591adb747f66385f
-folderAsset: yes
-DefaultImporter:
+guid: a8728824dd7e3554097f1c8f9deab3d9
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
