+++ Universe/Assets/Animations/Monsters/Boss/Boss.controller.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 8a840f51798911f4fb4b1d2c37bdcf35
+guid: 6fb2f98f634ce514a8eb447f96d15215
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 9100000
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
