+++ Universe/Assets/Animations/Monsters/Boss/BossAttack Animation.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: Boss Animation
+  m_Name: BossAttack Animation
   serializedVersion: 7
   m_Legacy: 0
   m_Compressed: 0
@@ -21,11 +21,9 @@ AnimationClip:
   - serializedVersion: 2
     curve:
     - time: 0
-      value: {fileID: 21300008, guid: 6f5c6e1922389384cb1736af0690439b, type: 3}
+      value: {fileID: 21300020, guid: 6f5c6e1922389384cb1736af0690439b, type: 3}
     - time: 0.16666667
-      value: {fileID: 21300010, guid: 6f5c6e1922389384cb1736af0690439b, type: 3}
-    - time: 0.33333334
-      value: {fileID: 21300012, guid: 6f5c6e1922389384cb1736af0690439b, type: 3}
+      value: {fileID: 21300022, guid: 6f5c6e1922389384cb1736af0690439b, type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
@@ -48,15 +46,14 @@ AnimationClip:
       isIntCurve: 0
       isSerializeReferenceCurve: 0
     pptrCurveMapping:
-    - {fileID: 21300008, guid: 6f5c6e1922389384cb1736af0690439b, type: 3}
-    - {fileID: 21300010, guid: 6f5c6e1922389384cb1736af0690439b, type: 3}
-    - {fileID: 21300012, guid: 6f5c6e1922389384cb1736af0690439b, type: 3}
+    - {fileID: 21300020, guid: 6f5c6e1922389384cb1736af0690439b, type: 3}
+    - {fileID: 21300022, guid: 6f5c6e1922389384cb1736af0690439b, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.35000002
+    m_StopTime: 0.18333334
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
