+++ Universe/Assets/Animations/Monsters/Boss/BossAttack Animation.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2364c73219f7cd0419a06741c9e115a5
+guid: de35c418290936e4ea904bf4ec47e842
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 7400000
