+++ Universe/Assets/Animations/Monsters/Boss/Idle01.controller.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 7c0c37c603a335b439a3223c20326163
-PrefabImporter:
+guid: ed3b2971b7edf264cb124a6069363491
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
