+++ Universe/Assets/Animations/Monsters/middleMonster/BeAttacked Animation.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: Boss Animation
+  m_Name: BeAttacked Animation
   serializedVersion: 7
   m_Legacy: 0
   m_Compressed: 0
@@ -21,17 +21,13 @@ AnimationClip:
   - serializedVersion: 2
     curve:
     - time: 0
-      value: {fileID: 21300000, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
+      value: {fileID: 21300000, guid: 1e86708f6739f9d45a11a727fa710193, type: 3}
     - time: 0.11666667
-      value: {fileID: 21300008, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
+      value: {fileID: 21300000, guid: 183f15fefcd558f4ab2841f369951d45, type: 3}
     - time: 0.23333333
-      value: {fileID: 21300010, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
-    - time: 0.36666667
-      value: {fileID: 21300002, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
-    - time: 0.5
-      value: {fileID: 21300006, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
-    - time: 0.6333333
-      value: {fileID: 21300010, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
+      value: {fileID: 21300000, guid: 86dd7dc47e8a22b46bb5c58159c15f04, type: 3}
+    - time: 0.35
+      value: {fileID: 21300000, guid: cc923659320ca1642901ea7336aff44f, type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
@@ -54,23 +50,21 @@ AnimationClip:
       isIntCurve: 0
       isSerializeReferenceCurve: 0
     pptrCurveMapping:
-    - {fileID: 21300000, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
-    - {fileID: 21300008, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
-    - {fileID: 21300010, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
-    - {fileID: 21300002, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
-    - {fileID: 21300006, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
-    - {fileID: 21300010, guid: 125a63fa69f337b48a5aad9eb6674b2d, type: 3}
+    - {fileID: 21300000, guid: 1e86708f6739f9d45a11a727fa710193, type: 3}
+    - {fileID: 21300000, guid: 183f15fefcd558f4ab2841f369951d45, type: 3}
+    - {fileID: 21300000, guid: 86dd7dc47e8a22b46bb5c58159c15f04, type: 3}
+    - {fileID: 21300000, guid: cc923659320ca1642901ea7336aff44f, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.65
+    m_StopTime: 0.36666667
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
     m_HasAdditiveReferencePose: 0
-    m_LoopTime: 1
+    m_LoopTime: 0
     m_LoopBlend: 0
     m_LoopBlendOrientation: 0
     m_LoopBlendPositionY: 0
