+++ Universe/Assets/Animations/Monsters/middleMonster/BeAttacked Animation.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2ed5e499d811aff4a98aa9048cc587d8
+guid: d27aca14e140b894d8c61d05dceeab04
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 7400000
