+++ Universe/Assets/DownlodeAssets/SunnyLand Artwork/Sprites/Enemies/Slimer-Idle/slimer-idle1 1.controller.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 3399d666c5b715a42a50113c88b8befe
-PrefabImporter:
+guid: 964b0f6b53c4a824f89d8b7c8e1586bc
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
