+++ Universe/Assets/Prefabs/Monsters/AngryMonster.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &2098115043172767626
+--- !u!1 &8979882756301850775
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,40 +8,55 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 2953983517094329412}
-  - component: {fileID: 5598865798110429180}
-  - component: {fileID: 6635319550817751587}
-  - component: {fileID: -2771329186174705741}
-  - component: {fileID: -6667366510351587002}
+  - component: {fileID: 3625859716634720692}
+  - component: {fileID: 5679734628702326129}
+  - component: {fileID: 1504916065729083531}
+  - component: {fileID: 414388159190797888}
+  - component: {fileID: 5094027220103768639}
+  - component: {fileID: 5364060467279347371}
+  - component: {fileID: -7229884197953683907}
   m_Layer: 0
-  m_Name: DieBoss
+  m_Name: AngryMonster
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &2953983517094329412
+--- !u!4 &3625859716634720692
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2098115043172767626}
+  m_GameObject: {fileID: 8979882756301850775}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1.2, y: 1.2, z: 1.2}
+  m_LocalScale: {x: 5, y: 5, z: 5}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!212 &5598865798110429180
+--- !u!114 &5679734628702326129
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8979882756301850775}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 18ee7a48a80c2274c85379376842c2f1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  BM_upgrade: {fileID: 1504916065729083531}
+--- !u!212 &1504916065729083531
 SpriteRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2098115043172767626}
+  m_GameObject: {fileID: 8979882756301850775}
   m_Enabled: 1
   m_CastShadows: 0
   m_ReceiveShadows: 0
@@ -76,28 +91,28 @@ SpriteRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 1
-  m_Sprite: {fileID: 21300036, guid: 6f5c6e1922389384cb1736af0690439b, type: 3}
+  m_Sprite: {fileID: 21300016, guid: 9d69ae98725365b4b94189120dc1b40a, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
-  m_Size: {x: 1, y: 1}
+  m_Size: {x: 0.16, y: 0.16}
   m_AdaptiveModeThreshold: 0.5
   m_SpriteTileMode: 0
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!95 &6635319550817751587
+--- !u!95 &414388159190797888
 Animator:
   serializedVersion: 5
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2098115043172767626}
+  m_GameObject: {fileID: 8979882756301850775}
   m_Enabled: 1
   m_Avatar: {fileID: 0}
-  m_Controller: {fileID: 9100000, guid: d304f27d765a5ea4e8c68f21861cab44, type: 2}
+  m_Controller: {fileID: 9100000, guid: 76981e935588a244296eaaf432e03c19, type: 2}
   m_CullingMode: 0
   m_UpdateMode: 0
   m_ApplyRootMotion: 0
@@ -108,13 +123,26 @@ Animator:
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorStateOnDisable: 0
   m_WriteDefaultValuesOnDisable: 0
---- !u!61 &-2771329186174705741
+--- !u!114 &5094027220103768639
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8979882756301850775}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ac4b046c3be510743b05fa4dcaba08a4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  BasicMoster_Rigidbody: {fileID: 0}
+--- !u!61 &5364060467279347371
 BoxCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2098115043172767626}
+  m_GameObject: {fileID: 8979882756301850775}
   m_Enabled: 1
   m_Density: 1
   m_Material: {fileID: 0}
@@ -140,27 +168,27 @@ BoxCollider2D:
   m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
-  m_Offset: {x: 0, y: 0}
+  m_Offset: {x: 0.00039321557, y: -0.02359324}
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 1, y: 1}
-    newSize: {x: 1, y: 1}
+    oldSize: {x: 0.16, y: 0.16}
+    newSize: {x: 0.16, y: 0.16}
     adaptiveTilingThreshold: 0.5
     drawMode: 0
     adaptiveTiling: 0
   m_AutoTiling: 0
   serializedVersion: 2
-  m_Size: {x: 1, y: 1}
+  m_Size: {x: 0.14191183, y: 0.11281352}
   m_EdgeRadius: 0
---- !u!50 &-6667366510351587002
+--- !u!50 &-7229884197953683907
 Rigidbody2D:
   serializedVersion: 4
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2098115043172767626}
+  m_GameObject: {fileID: 8979882756301850775}
   m_BodyType: 0
   m_Simulated: 1
   m_UseFullKinematicContacts: 0
