+++ Universe/Assets/Prefabs/Monsters/Boss.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5525d6f5ce7432d429c453353e28938c
+guid: 336ca2254efa9c04a8d684b415e9538e
 PrefabImporter:
   externalObjects: {}
   userData:
