+++ Universe/Assets/Scenes/WoojeongScene.unity.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 2ed5e499d811aff4a98aa9048cc587d8
-NativeFormatImporter:
+guid: 2cda990e2423bbf4892e6590ba056729
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
