+++ Universe/Assets/Scripts/Monster.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: a32abec935b40f140898b8ad9f6c8421
-PrefabImporter:
+guid: 29671139786952c41baf10dc8b02bda9
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
