+++ Universe/Assets/Scripts/Monster/BasicMonster.cs.meta
