+++ Universe/Assets/Scripts/Monster/BulletForward.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4aaaeffb29481b4ab6e7a58cd68ec1c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
